feat: step-by-step execution
Now, the list of BrickCodes is executed one by one in order, and it's represented by the red arrow on the side of the corresponding brick. To start or stop de execution of the code, press the "A" button on the right controller (quest 3). If you stop the execution, the current code that's running will finish and only after that, the execution will stop, to prevent errors. You can resume the execution by pressing the "A" button again.