feat: add bricks
pressing the grip trigger on the right controller brings up a hovering menu on top of the controller that shows a lsit of possible bricks to add. Using the left crontroller the user can select and press X on some brick to add them to the list, so the robot can perform the task!